# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Categorías de Productos",
        "description": "API para gestionar categorías de productos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categoria_emprendimiento": {
            "get": {
                "tags": [
                    "Categorías de Emprendimiento"
                ],
                "summary": "Obtener todas las categorías de emprendimiento",
                "operationId": "5b364f73395d6cb48600f12ae32b94a0",
                "responses": {
                    "200": {
                        "description": "Lista de categorías obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Tecnología"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Emprendimientos tecnológicos"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No hay categorías registradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "no hay categorias registradas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías de Emprendimiento"
                ],
                "summary": "Crear una nueva categoría de emprendimiento",
                "operationId": "fe57841540d3c13d11fc560c8c0937fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado",
                                    "nombre"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "activo"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Tecnología"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Emprendimientos tecnológicos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Categoría creada con éxito"
                                        },
                                        "categoria": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "estado": {
                                                    "type": "string",
                                                    "example": "activo"
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Tecnología"
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Emprendimientos tecnológicos"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de los datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de los datos"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear categoría",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear categoría"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}